CreateAuthors page and fixing include

+++ Private/CRUD/delete.php
@@ -1,7 +1,17 @@
 <?php
 function delete_author(int $author_id)
 {
-    include("../Misc/config.php");
+    $serverName = "localhost";
+    $userName = "root";
+    $password = "";
+    $dbName = "alexandria";
+
+    if (!$dbConn = mysqli_connect($serverName, $userName, $password, $dbName)) {
+        die("Could not connect to db<br>" . mysqli_connect_error());
+    }
+    if (!mysqli_select_db($dbConn, $dbName)) {
+        die("Could not select db<br>" . mysqli_connect_error());
+    }
     $sql = "DELETE FROM authors WHERE AuthorId=$author_id";
     $result = mysqli_query($dbConn, $sql);
     if (!$result) {
@@ -10,7 +20,17 @@ function delete_author(int $author_id)
 }
 function delete_genre(int $genre_id)
 {
-    include("../Misc/config.php");
+    $serverName = "localhost";
+    $userName = "root";
+    $password = "";
+    $dbName = "alexandria";
+
+    if (!$dbConn = mysqli_connect($serverName, $userName, $password, $dbName)) {
+        die("Could not connect to db<br>" . mysqli_connect_error());
+    }
+    if (!mysqli_select_db($dbConn, $dbName)) {
+        die("Could not select db<br>" . mysqli_connect_error());
+    }
     $sql = "DELETE FROM genres WHERE GenreId=$genre_id";
     $result = mysqli_query($dbConn, $sql);
     if (!$result) {
@@ -19,7 +39,17 @@ function delete_genre(int $genre_id)
 }
 function delete_position(int $position_id)
 {
-    include("../Misc/config.php");
+    $serverName = "localhost";
+    $userName = "root";
+    $password = "";
+    $dbName = "alexandria";
+
+    if (!$dbConn = mysqli_connect($serverName, $userName, $password, $dbName)) {
+        die("Could not connect to db<br>" . mysqli_connect_error());
+    }
+    if (!mysqli_select_db($dbConn, $dbName)) {
+        die("Could not select db<br>" . mysqli_connect_error());
+    }
     $sql = "DELETE FROM positions WHERE PositionId=$position_id";
     $result = mysqli_query($dbConn, $sql);
     if (!$result) {
@@ -28,7 +58,17 @@ function delete_position(int $position_id)
 }
 function delete_borrow(int $borrow_id)
 {
-    include("../Misc/config.php");
+    $serverName = "localhost";
+    $userName = "root";
+    $password = "";
+    $dbName = "alexandria";
+
+    if (!$dbConn = mysqli_connect($serverName, $userName, $password, $dbName)) {
+        die("Could not connect to db<br>" . mysqli_connect_error());
+    }
+    if (!mysqli_select_db($dbConn, $dbName)) {
+        die("Could not select db<br>" . mysqli_connect_error());
+    }
     $sql = "DELETE FROM borrows WHERE BorrowId=$borrow_id";
     $result = mysqli_query($dbConn, $sql);
     if (!$result) {
@@ -37,7 +77,17 @@ function delete_borrow(int $borrow_id)
 }
 function delete_client(int $client_id)
 {
-    include("../Misc/config.php");
+    $serverName = "localhost";
+    $userName = "root";
+    $password = "";
+    $dbName = "alexandria";
+
+    if (!$dbConn = mysqli_connect($serverName, $userName, $password, $dbName)) {
+        die("Could not connect to db<br>" . mysqli_connect_error());
+    }
+    if (!mysqli_select_db($dbConn, $dbName)) {
+        die("Could not select db<br>" . mysqli_connect_error());
+    }
     $sql = "DELETE FROM clients WHERE ClientId=$client_id";
     $result = mysqli_query($dbConn, $sql);
     if (!$result) {
@@ -46,7 +96,17 @@ function delete_client(int $client_id)
 }
 function delete_employee(int $employee_id)
 {
-    include("../Misc/config.php");
+    $serverName = "localhost";
+    $userName = "root";
+    $password = "";
+    $dbName = "alexandria";
+
+    if (!$dbConn = mysqli_connect($serverName, $userName, $password, $dbName)) {
+        die("Could not connect to db<br>" . mysqli_connect_error());
+    }
+    if (!mysqli_select_db($dbConn, $dbName)) {
+        die("Could not select db<br>" . mysqli_connect_error());
+    }
     $sql = "DELETE FROM employees WHERE EmployeeId=$employee_id";
     $result = mysqli_query($dbConn, $sql);
     if (!$result) {
@@ -55,7 +115,17 @@ function delete_employee(int $employee_id)
 }
 function delete_book_authors(int $book_id, int $author_id)
 {
-    include("../Misc/config.php");
+    $serverName = "localhost";
+    $userName = "root";
+    $password = "";
+    $dbName = "alexandria";
+
+    if (!$dbConn = mysqli_connect($serverName, $userName, $password, $dbName)) {
+        die("Could not connect to db<br>" . mysqli_connect_error());
+    }
+    if (!mysqli_select_db($dbConn, $dbName)) {
+        die("Could not select db<br>" . mysqli_connect_error());
+    }
     $sql = "DELETE FROM book_authors WHERE BookId=$book_id AND AuthorId=$author_id";
     $result = mysqli_query($dbConn, $sql);
     if (!$result) {
@@ -64,7 +134,17 @@ function delete_book_authors(int $book_id, int $author_id)
 }
 function delete_book(int $book_id)
 {
-    include("../Misc/config.php");
+    $serverName = "localhost";
+    $userName = "root";
+    $password = "";
+    $dbName = "alexandria";
+
+    if (!$dbConn = mysqli_connect($serverName, $userName, $password, $dbName)) {
+        die("Could not connect to db<br>" . mysqli_connect_error());
+    }
+    if (!mysqli_select_db($dbConn, $dbName)) {
+        die("Could not select db<br>" . mysqli_connect_error());
+    }
     $sql = "DELETE FROM books WHERE BookId=$book_id";
     $result = mysqli_query($dbConn, $sql);
     if (!$result) {
